Split claim runway snapshot tests
