+++ crates/oasis7/src/viewer/runtime_live/tests.rs
@@ -18,6 +18,7 @@ mod background_play;
 mod chain_sync;
 pub(super) use chain_sync::TestChainStatusServer;
 mod chain_sync_feedback;
+mod claim_runway;
 mod industrial_progression;
 mod prompt_control;
 mod provider_settings;
