+++ crates/oasis7/src/viewer/runtime_live/tests/snapshot_progress.rs
@@ -864,66 +864,6 @@ fn compat_snapshot_exposes_slot_1_auto_funding_from_dedicated_pool() {
 }
 
 #[test]
-fn compat_snapshot_quotes_complete_epoch_runway_and_advisory_threshold() {
-    for (balance, expected_after, expected_runway, expected_warning, expected_action) in [
-        (375, 50, 2, false, "claim_now_route_fit"),
-        (374, 49, 1, true, "wait_or_fund_first"),
-    ] {
-        let mut server = ViewerRuntimeLiveServer::new(ViewerRuntimeLiveServerConfig::new(
-            WorldScenario::Minimal,
-        ))
-        .expect("runtime server");
-        let primary_agent_id = server
-            .world
-            .state()
-            .agents
-            .keys()
-            .next()
-            .cloned()
-            .expect("primary agent");
-        bind_agent_for_snapshot(&mut server, primary_agent_id.as_str());
-        server
-            .world
-            .set_governance_execution_policy(crate::runtime::GovernanceExecutionPolicy {
-                epoch_length_ticks: 1,
-                ..crate::runtime::GovernanceExecutionPolicy::default()
-            })
-            .expect("set governance policy");
-        server
-            .world
-            .set_main_token_supply(crate::runtime::MainTokenSupplyState {
-                total_supply: balance,
-                circulating_supply: balance,
-                ..crate::runtime::MainTokenSupplyState::default()
-            });
-        server
-            .world
-            .set_main_token_account_balance(primary_agent_id.as_str(), balance, 0)
-            .expect("seed main token balance");
-
-        let snapshot = server.compat_snapshot(Some(SNAPSHOT_PLAYER_ID));
-        let claim = snapshot
-            .player_gameplay
-            .as_ref()
-            .and_then(|gameplay| gameplay.agent_claim.as_ref())
-            .expect("player agent claim snapshot");
-        let quote = claim.next_claim_quote.as_ref().expect("next claim quote");
-        assert_eq!(quote.total_upfront_amount, 325);
-        assert_eq!(quote.upkeep_per_epoch, 25);
-        assert_eq!(quote.grace_epochs, 2);
-        assert_eq!(quote.eligible_balance_after, expected_after);
-        assert_eq!(quote.upkeep_runway_epochs, expected_runway);
-        assert_eq!(quote.next_upkeep_due_epoch, Some(claim.current_epoch + 1));
-        assert_eq!(quote.low_runway_warning, expected_warning);
-        assert_eq!(
-            quote.recommended_claim_action.as_deref(),
-            Some(expected_action)
-        );
-        assert_eq!(quote.blocked_reason, None, "warning must remain advisory");
-    }
-}
-
-#[test]
 fn compat_snapshot_promotes_to_post_onboarding_after_control_feedback() {
     let _guard = lock_test_llm_env();
     let mut server = ViewerRuntimeLiveServer::new(
